+++ src/ddzy/utility/algorithm/es6-achieve/index.ts
@@ -10,6 +10,7 @@ import { _reflect } from './reflect/index';
 import { _map } from './map/index';
 import { _reduce } from './reduce';
 import { _filter } from './filter';
+import { _every } from './every/index';
 
 
 export const ES6Achieve = {
@@ -18,4 +19,5 @@ export const ES6Achieve = {
   _map,
   _reduce,
   _filter,
+  _every,
 };