+++ src/ddzy/CHANGELOG.md
@@ -196,3 +196,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 - `utility/algorithm/es6achieve`
   - 新增`_filter`
+
+## ✍ [2019-7-14]
+
+### Added
+
+- utility/algorithm/es6achieve
+  - 新增`_every`